+++ src/ui/device.win.controls/Views/DisplayViewSettings.cs
@@ -41,7 +41,9 @@ public class DisplayViewSettings : cc.isr.Json.AppSettings.Settings.SettingsBase
             SectionName = declaringType.Name
         };
 
-        ti.ReadSettings( declaringType, ".Settings", System.Diagnostics.Debugger.IsAttached, System.Diagnostics.Debugger.IsAttached );
+        bool overwrite = Json.AppSettings.Models.AppSettingsScribe.IsDebuggingOrTesting();
+
+        ti.ReadSettings( declaringType, ".Settings", overwrite, overwrite );
 
         return ti;
     }