+++ src/ui/win.controls.mstest/Tests/AllSettings.cs
@@ -43,12 +43,12 @@ public class AllSettings : cc.isr.Json.AppSettings.Settings.SettingsContainerBas
     /// <param name="settingsFileSuffix">   (Optional) [.settings] The suffix of the assembly settings file,
     ///                                     e.g., '.settings' in 'cc.isr.VI.Tsp.K2600.Device.MSTest.Session.json'
     ///                                     where cc.isr.VI.Tsp.K2600.Device.MSTest is the assembly name.<,/param>
-    /// <param name="overrideAllUsersFile"> (Optional) [false] True to override all users settings file. </param>
-    /// <param name="overrideThisUserFile"> (Optional) [false] True to override this user settings file. </param>
+    /// <param name="overwriteAllUsersFile"> (Optional) [false] True to over-write all users settings file. </param>
+    /// <param name="overwriteThisUserFile"> (Optional) [false] True to over-write this user settings file. </param>
     public override void ReadSettings( Type declaringType, string settingsFileSuffix = ".Settings",
-        bool overrideAllUsersFile = false, bool overrideThisUserFile = false )
+        bool overwriteAllUsersFile = false, bool overwriteThisUserFile = false )
     {
-        base.ReadSettings( declaringType, settingsFileSuffix, overrideAllUsersFile, overrideThisUserFile );
+        base.ReadSettings( declaringType, settingsFileSuffix, overwriteAllUsersFile, overwriteThisUserFile );
     }
 
     #endregion