+++ src/vi/k2600/k2600.mstest/Tests/Settings/AllSettings.cs
@@ -61,7 +61,9 @@ public class AllSettings : Json.AppSettings.Settings.SettingsContainerBase
 
         AllSettings ti = new();
 
-        ti.ReadSettings( declaringType, overrideAllUsersFile: true, overrideThisUserFile: true );
+        bool overwrite = Json.AppSettings.Models.AppSettingsScribe.IsDebuggingOrTesting();
+
+        ti.ReadSettings( declaringType, overwriteAllUsersFile: overwrite, overwriteThisUserFile: overwrite );
 
         return ti;
     }
@@ -85,12 +87,12 @@ public class AllSettings : Json.AppSettings.Settings.SettingsContainerBase
     ///                                     '.session' in
     ///                                     'cc.isr.VI.Tsp.K2600.Device.MSTest.Session.json' where
     ///                                     cc.isr.VI.Tsp.K2600.Device.MSTest is the assembly name. </param>
-    /// <param name="overrideAllUsersFile"> (Optional) [false] True to override all users settings file. </param>
-    /// <param name="overrideThisUserFile"> (Optional) [false] True to override this user settings file. </param>
+    /// <param name="overwriteAllUsersFile"> (Optional) [false] True to over-write all users settings file. </param>
+    /// <param name="overwriteThisUserFile"> (Optional) [false] True to over-write this user settings file. </param>
     public override void ReadSettings( Type declaringType, string settingsFileSuffix = ".Settings",
-        bool overrideAllUsersFile = false, bool overrideThisUserFile = false )
+        bool overwriteAllUsersFile = false, bool overwriteThisUserFile = false )
     {
-        base.ReadSettings( declaringType, settingsFileSuffix, overrideAllUsersFile, overrideThisUserFile );
+        base.ReadSettings( declaringType, settingsFileSuffix, overwriteAllUsersFile, overwriteThisUserFile );
     }
 
     #endregion