+++ src/vi/k2600/k2600.mstest/Tests/Status/DeviceStatusOnlyTests.cs
@@ -52,7 +52,7 @@ public class DeviceStatusOnlyTests : Device.Tests.Base.DeviceStatusOnlyTests
         this.Device = K2600Device.Create();
         Assert.IsNotNull( this.Device );
         Assert.IsNotNull( this.Device.Session );
-        this.Device.Session.ReadSettings( this.GetType(), overrideAllUsersFile: true, overrideThisUserFile: true );
+        this.Device.Session.ReadSettings( this.GetType(), overwriteAllUsersFile: true, overwriteThisUserFile: true );
         this.VisaSessionBase = this.Device;
         base.InitializeBeforeEachTest();
     }