+++ changelog.md
@@ -2,9 +2,13 @@
 Notable changes to the `cc.isr.vi` solution are documented in this file using the 
 [Keep a Changelog] style. The dates specified are in coordinated universal time (UTC).
 
-[8.1.9451]: https://www.github.com/atecoder/dn.vi.ivi
+[8.1.9453]: https://www.github.com/atecoder/dn.vi.ivi
 
 Current IVI Visa Compatibility: 8.0.1.9391
+
+## [8.1.9453] - 2025-11-18
+- Transition to .SLNX solution file.
+- Overwrite settings if debugging or testing.
 
 ## [8.1.9451] - 2025-11-16
 - MS Test: Initialize settings files.